newcoop_app/requirements.txt
django-haystack
django
psycopg2
django-allauth
whoosh
